Remove secret pages-page
