--- front/app/containers/Admin/pages/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export interface Props {}
-
-interface State {}
-
-export default class PagesDashboard extends React.PureComponent<Props, State> {
-  render() {
-    return <>{this.props.children}</>;
-  }
-}
