--- front/app/containers/Admin/pages/messages.ts
@@ -1,24 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  listTitle: {
-    id: 'app.containers.AdminPage.PagesPage.listTitle',
-    defaultMessage: 'Pages',
-  },
-  pageDeletionConfirmation: {
-    id: 'app.containers.AdminPage.PagesPage.pageDeletionConfirmation',
-    defaultMessage: 'Are you sure you want to delete this page?',
-  },
-  showButtonLabel: {
-    id: 'app.containers.AdminPage.PagesPage.showButtonLabel',
-    defaultMessage: 'View',
-  },
-  deleteButtonLabel: {
-    id: 'app.containers.AdminPage.PagesPage.deleteButtonLabel',
-    defaultMessage: 'Delete',
-  },
-  editButtonLabel: {
-    id: 'app.containers.AdminPage.PagesPage.editButtonLabel',
-    defaultMessage: 'Edit',
-  },
-});
