+++ front/app/modules/commercial/customizable_navbar/admin/containers/EditNavbarItemForm/utils.ts
@@ -1,8 +1,7 @@
-import { INavbarItem } from 'services/navbar';
+import { INavbarItem, MAX_TITLE_LENGTH } from 'services/navbar';
 import { FormValues } from '../../components/NavbarItemForm';
 import { INavbarItemUpdate } from '../../../services/navbar';
 import { truncateMultiloc } from 'utils/textUtils';
-import { MAX_TITLE_LENGTH } from 'services/navbar';
 
 export const getInitialFormValues = (navbarItem: INavbarItem): FormValues => ({
   nav_bar_item_title_multiloc: truncateMultiloc(
