+++ front/app/modules/commercial/customizable_navbar/admin/containers/VisibleNavbarItemList/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 // services
 import { reorderNavbarItem, removeNavbarItem } from '../../../services/navbar';
-import { deletePage } from 'services/pages';
+import { deletePage } from '../../../services/pages';
 import { INavbarItem, getNavbarItemSlug } from 'services/navbar';
 
 // components
