+++ front/app/modules/commercial/customizable_navbar/hooks/useNavbarItem.ts
@@ -26,7 +26,7 @@ export default function useNavbarItem({ navbarItemId }: IParams) {
     );
 
     return () => subscription.unsubscribe();
-  }, []);
+  }, [navbarItemId]);
 
   return navbarItem;
 }
