+++ front/app/modules/commercial/customizable_navbar/services/pages.ts
@@ -3,6 +3,12 @@ import { apiEndpoint, IPage } from 'services/pages';
 import { apiEndpoint as navbarEndpoint } from 'services/navbar';
 import { Multiloc } from 'typings';
 
+interface IPageCreate {
+  title_multiloc: Multiloc;
+  body_multiloc: Multiloc;
+  slug?: string;
+}
+
 export interface IPageUpdate {
   title_multiloc?: Multiloc;
   body_multiloc?: Multiloc;
@@ -10,6 +16,10 @@ export interface IPageUpdate {
   nav_bar_item_attributes?: {
     title_multiloc?: Multiloc;
   };
+}
+
+export function createPage(pageData: IPageCreate) {
+  return streams.add<IPage>(`${apiEndpoint}`, pageData);
 }
 
 export async function updatePage(pageId: string, pageData: IPageUpdate) {
@@ -22,6 +32,12 @@ export async function updatePage(pageId: string, pageData: IPageUpdate) {
   await streams.fetchAllWith({
     apiEndpoint: [navbarEndpoint],
   });
+
+  return response;
+}
+export async function deletePage(pageId: string) {
+  const response = await streams.delete(`${apiEndpoint}/${pageId}`, pageId);
+  await streams.fetchAllWith({ apiEndpoint: [navbarEndpoint] });
 
   return response;
 }
